tidied up the Rollup build process and split functions into various different files to improve readability
