+++ scripts/rollup/results.json
@@ -2,116 +2,116 @@
   "branch": "rollup",
   "bundleSizes": {
     "react.development.js (UMD_DEV)": {
-      "size": 116398,
-      "gzip": 29597
+      "size": 116402,
+      "gzip": 29598
     },
     "react.production.min.js (UMD_PROD)": {
-      "size": 13980,
-      "gzip": 5170
+      "size": 13719,
+      "gzip": 5080
     },
     "react-dom.development.js (UMD_DEV)": {
-      "size": 565398,
-      "gzip": 130384
+      "size": 565236,
+      "gzip": 130605
     },
     "react-dom.production.min.js (UMD_PROD)": {
-      "size": 115831,
-      "gzip": 36755
+      "size": 120184,
+      "gzip": 37868
     },
     "react-dom-server.development.js (UMD_DEV)": {
-      "size": 477493,
-      "gzip": 115611
+      "size": 477786,
+      "gzip": 115723
     },
     "react-dom-server.production.min.js (UMD_PROD)": {
-      "size": 106795,
-      "gzip": 33052
+      "size": 106410,
+      "gzip": 33019
     },
     "react-art.development.js (UMD_DEV)": {
-      "size": 341671,
-      "gzip": 76162
+      "size": 341133,
+      "gzip": 76276
     },
     "react-art.production.min.js (UMD_PROD)": {
-      "size": 91056,
-      "gzip": 28078
+      "size": 94939,
+      "gzip": 28977
     },
     "react.development.js (NODE_DEV)": {
-      "size": 109412,
-      "gzip": 27553
+      "size": 109416,
+      "gzip": 27554
     },
     "react.production.min.js (NODE_PROD)": {
-      "size": 12742,
-      "gzip": 4693
+      "size": 12615,
+      "gzip": 4659
     },
     "React-dev.js (FB_DEV)": {
-      "size": 110897,
+      "size": 110901,
       "gzip": 28115
     },
     "React-prod.js (FB_PROD)": {
-      "size": 56438,
-      "gzip": 14394
+      "size": 56205,
+      "gzip": 14328
     },
     "ReactDOMStack-dev.js (FB_DEV)": {
-      "size": 523961,
-      "gzip": 125060
+      "size": 523977,
+      "gzip": 125062
     },
     "ReactDOMStack-prod.js (FB_PROD)": {
-      "size": 355546,
-      "gzip": 85701
+      "size": 351653,
+      "gzip": 84412
     },
     "react-dom.development.js (NODE_DEV)": {
-      "size": 544119,
-      "gzip": 125425
+      "size": 543610,
+      "gzip": 125563
     },
     "react-dom.production.min.js (NODE_PROD)": {
-      "size": 112912,
-      "gzip": 35728
+      "size": 116863,
+      "gzip": 36727
     },
     "ReactDOMFiber-dev.js (FB_DEV)": {
-      "size": 798172,
-      "gzip": 184345
+      "size": 798188,
+      "gzip": 184347
     },
     "ReactDOMFiber-prod.js (FB_PROD)": {
-      "size": 412153,
-      "gzip": 94702
+      "size": 407208,
+      "gzip": 93465
     },
     "react-dom-server.development.js (NODE_DEV)": {
-      "size": 447241,
-      "gzip": 108016
+      "size": 447187,
+      "gzip": 108026
     },
     "react-dom-server.production.min.js (NODE_PROD)": {
-      "size": 101538,
-      "gzip": 31190
+      "size": 101265,
+      "gzip": 31247
     },
     "ReactDOMServerStack-dev.js (FB_DEV)": {
-      "size": 445922,
-      "gzip": 107879
+      "size": 445938,
+      "gzip": 107881
     },
     "ReactDOMServerStack-prod.js (FB_PROD)": {
-      "size": 336779,
-      "gzip": 81495
+      "size": 332920,
+      "gzip": 80268
     },
     "ReactARTStack-dev.js (FB_DEV)": {
-      "size": 494733,
-      "gzip": 118996
+      "size": 494749,
+      "gzip": 118999
     },
     "ReactARTStack-prod.js (FB_PROD)": {
-      "size": 368975,
-      "gzip": 88944
+      "size": 365061,
+      "gzip": 87642
     },
     "react-art.development.js (NODE_DEV)": {
-      "size": 266906,
-      "gzip": 57091
+      "size": 266368,
+      "gzip": 57200
     },
     "react-art.production.min.js (NODE_PROD)": {
-      "size": 52925,
-      "gzip": 16250
+      "size": 56517,
+      "gzip": 17109
     },
     "ReactARTFiber-dev.js (FB_DEV)": {
-      "size": 265539,
+      "size": 265543,
       "gzip": 57024
     },
     "ReactARTFiber-prod.js (FB_PROD)": {
-      "size": 207824,
-      "gzip": 43477
+      "size": 205194,
+      "gzip": 43039
     },
     "ReactNativeStack.js (RN)": {
       "size": 233993,
@@ -122,20 +122,20 @@
       "gzip": 84001
     },
     "ReactTestRendererFiber-dev.js (FB_DEV)": {
-      "size": 263430,
+      "size": 263434,
       "gzip": 55987
     },
     "ReactTestRendererStack-dev.js (FB_DEV)": {
-      "size": 155200,
-      "gzip": 35680
+      "size": 155204,
+      "gzip": 35682
     },
     "react-noop-renderer.development.js (NODE_DEV)": {
-      "size": 255932,
-      "gzip": 53845
+      "size": 255429,
+      "gzip": 53949
     },
     "react-test-renderer.development.js (NODE_DEV)": {
-      "size": 264826,
-      "gzip": 56049
+      "size": 264268,
+      "gzip": 56169
     }
   }
 }
